+++ CMakeLists.txt
@@ -20,7 +20,7 @@ include_directories(src/inference)
 
 add_library(${project_name}-lib
         src/inference/tokenizer.cpp
-        src/inference/tokeniser.h)
+        src/inference/tokeniszer.hpp)
 
 # Executable to add that will be compiled and run
 add_executable(print_vocab
