+++ src/inference/tokenizer.cpp
@@ -1,4 +1,4 @@
-#include "tokeniser.h"
+#include "tokenizer.h"
 #include <codecvt>
 #include <locale>
 #include <cmath>
